Rename FsClient::getMetadata -> FsClient::inspect, add FsClient::inspectAll

+++ cloud-fs/src/main/java/io/activej/remotefs/FsClient.java
@@ -25,15 +25,13 @@ import io.activej.promise.Promises;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
 import static io.activej.common.collection.CollectionUtils.map;
 import static io.activej.remotefs.RemoteFsUtils.escapeGlob;
+import static java.util.Collections.emptyMap;
 
 /**
  * This interface represents a simple filesystem client with upload, download, copy, delete and list operations.
@@ -133,6 +131,8 @@ public interface FsClient {
 	 * @param sourceToTarget source files to target files mapping
 	 */
 	default Promise<Void> copyAll(Map<String, String> sourceToTarget) {
+		if (sourceToTarget.isEmpty()) return Promise.complete();
+
 		return Promises.all(sourceToTarget.entrySet().stream()
 				.map(entry -> copy(entry.getKey(), entry.getValue())));
 	}
@@ -168,6 +168,8 @@ public interface FsClient {
 	 * @param sourceToTarget source files to target files mapping
 	 */
 	default Promise<Void> moveAll(Map<String, String> sourceToTarget) {
+		if (sourceToTarget.isEmpty()) return Promise.complete();
+
 		return copyAll(sourceToTarget)
 				.then(() -> deleteAll(sourceToTarget.keySet()));
 	}
@@ -188,9 +190,26 @@ public interface FsClient {
 	 * @param name name of a file to fetch its metadata.
 	 * @return promise of file description or <code>null</code>
 	 */
-	default Promise<@Nullable FileMetadata> getMetadata(@NotNull String name) {
+	default Promise<@Nullable FileMetadata> inspect(@NotNull String name) {
 		return list(escapeGlob(name))
 				.map(list -> list.isEmpty() ? null : list.get(0));
+	}
+
+	/**
+	 * Shortcut to get {@link FileMetadata metadata} of multiple files.
+	 *
+	 * @param names names of files to fetch metadata for.
+	 * @return map of filenames to their corresponding {@link FileMetadata metadata}.
+	 * If there is no such file, a value is <code>null</code>
+	 */
+	default Promise<Map<String, @Nullable FileMetadata>> inspectAll(@NotNull List<String> names) {
+		if (names.isEmpty()) return Promise.of(emptyMap());
+
+		Map<String, FileMetadata> result = new HashMap<>();
+		return Promises.all(names.stream()
+				.map(name -> inspect(name)
+						.whenResult(metadata -> result.put(name, metadata))))
+				.map($ -> result);
 	}
 
 	/**
